Clean up script
Address #10. Tests pass.

Clean up script by using new Variables and Equations classes. Change
necessary to remove a lot of the details from the script to make it
easier to concentrate on the iteration flow. Using the Variables and
Equations class removes a lot of the setup code leaving only the
iteration loops in the script.

 * move import tables to top of imports in tools.py

 * move DistanceVariableNonUniform to variables.py

 * remove Parameters class

 * replace MeshBuilder class with build_mesh function

 * move calc_current function into Variables class

 * move all parameters into JSON file

 * move variable creation to Variables class

 * move equation creation to Equations class

 * move other subsections of script to Equations class