Implement equation for applied potential
Address #10

Applied potential variables should be changed by an equation rather
than Variable update so the updates are controlled as part of the
sweep. The updates could occur multiple times in a sweep otherwise.

The equation building methods in `Equations` have been moved to
classes as this is more OO.

 * moved Equations' class methods into separate Equation classes

 * starting to implement AppliedPotentialEquation

 * `print_data` uses `1.2e` formatting for floats

 * `Variables` builds intermediate variables required by the
   `AppliedPotentialEquation`