Constant current is fully implemented
Fix #10. Tests pass.

Constant current is fully implemented. The size of the correction to
the `appliedPotential` variables is limited to maintain stability.

Tests have been refactored in a major way. A context manager is used
to setup and teardown to enable `import annular` to run like a
script. Unfortunately, further refactoring of the code will make this
unnecessary.

 * constant current now works

 * the size of the correction to `appliedPotential` is limited to
   preserve stability

 * create `ConstantCurrentEquations`

 * add self._requires to `AppliedPotentialVariable`

 * add `constant_current` parameter

 * add test for constant current

 * major refactor of test cases