New `ExtemeFill2DMesh` class
Address #16

If the mesh is changed from an annular representation to a via, the
`build_mesh` raises an exception. The mesh build needs to be fixed so
that it works for annular and via shapes. As a precursor to this the
mesh functions were refactored into a class to make future subclassing
easier.

 * include sumatra setup instructions in README.md

 * `import scipy` and `import scipy.optimize` as this stopped working

 * move mesh building functions to one ExtremeFill2DMesh class