Hemispherical boundary layer shape
Fix #17

Enable hemispherical boundary layer shape as gross approximation to the
influence that the feature might have on the boundary layer shape. The
cupric transient term coefficient is set to a very small value in the
boundary region. The transient term for the suppressor has not been
changed yet, but may need to be.